Update README.md
Welcome to the keralalotteries family. Here we will be updating latest kerala lottery result. We are also providing a detailed guide on claiming the winning amount of the lottery ticket. Today, you can check the Kerala Lottery Result from this page before any other website. We provide the fastest updates for the lottery draws around the country. Always remember that the lottery system is risky and purely depends on luck.

Take a look at our website www.keralalotteriesresult.in